requirements.txt
streamlit
pandas
scikit-learn
matplotlib
python-dotenv
langchain
openai
